refactor: enhance workspace and project wrapper modularity
