+++ web/app/[workspaceSlug]/(projects)/projects/(detail)/layout.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { ReactNode } from "react";
-// layouts
-import { ProjectAuthWrapper } from "@/layouts/auth-layout";
+// plane web layouts
+import { ProjectAuthWrapper } from "@/plane-web/layouts/project-wrapper";
 
 const ProjectDetailLayout = ({ children }: { children: ReactNode }) => (
   <ProjectAuthWrapper>{children}</ProjectAuthWrapper>
