+++ web/ee/layouts/project-wrapper.tsx
@@ -1,0 +1,1 @@
+export * from "ce/layouts/project-wrapper";
