+++ web/ee/layouts/workspace-wrapper.tsx
@@ -1,0 +1,1 @@
+export * from "ce/layouts/workspace-wrapper";
